+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Unreleased
+----------
+- Moved `Custom` event variant from `gui::Event` into `gui::UiEvent` and
+  renamed former `gui::UiEvent::Custom` into `gui::UiEvent::Directed`
+
+
 0.1.1
 -----
 - Added link to `docs.rs` based documentation to README